+++ src/modules/promotions/promotions.controller.ts
@@ -1,4 +1,7 @@
+import type { I18nTranslations } from '@types';
+
 import {
+	BadRequestException,
 	Controller,
 	Delete,
 	Get,
@@ -23,7 +26,9 @@ import {
 	ApiTags,
 	ApiUnauthorizedResponse,
 } from '@nestjs/swagger';
+import { I18nService } from 'nestjs-i18n';
 
+import { Errors } from '@i18n';
 import { GuardPermissions } from '@modules/auth/decorators/permissions.decorator';
 import { PermissionGuard } from '@modules/auth/guards/permission.guard';
 import { FilesService } from '@modules/files/files.service';
@@ -38,7 +43,11 @@ import { BaseUserResponseDTO } from '../users/dto/base-user.dto';
 @UseGuards(AuthGuard('jwt'))
 @ApiBearerAuth()
 export class PromotionsController {
-	constructor(private readonly promotionsService: PromotionsService, private readonly filesService: FilesService) {}
+	constructor(
+		private readonly promotionsService: PromotionsService,
+		private readonly filesService: FilesService,
+		private readonly i18n: I18nService<I18nTranslations>,
+	) {}
 
 	@Get()
 	@UseGuards(PermissionGuard)
@@ -73,6 +82,11 @@ export class PromotionsController {
 	})
 	@UseInterceptors(FileInterceptor('file'))
 	async editLogo(@UploadedFile() file: Express.Multer.File, @Param('number') number: number) {
+		if (!file) throw new BadRequestException(Errors.File.NotProvided({ i18n: this.i18n }));
+
+		if (typeof number !== 'number' && parseInt(number, 10) != number)
+			throw new BadRequestException(Errors.Generic.FieldInvalid({ i18n: this.i18n, type: Number, field: 'number' }));
+
 		return this.promotionsService.updateLogo(number, file);
 	}
 
@@ -85,6 +99,9 @@ export class PromotionsController {
 	@ApiUnauthorizedResponse({ description: 'Insufficient permission' })
 	@ApiNotFoundResponse({ description: 'Promotion not found or promotion has no logo' })
 	async getLogo(@Param('number') number: number) {
+		if (typeof number !== 'number' && parseInt(number, 10) != number)
+			throw new BadRequestException(Errors.Generic.FieldInvalid({ i18n: this.i18n, type: Number, field: 'number' }));
+
 		const logo = await this.promotionsService.getLogo(number);
 		return new StreamableFile(this.filesService.toReadable(logo));
 	}
@@ -98,6 +115,9 @@ export class PromotionsController {
 	@ApiOkResponse({ type: Promotion })
 	@ApiUnauthorizedResponse({ description: 'Insufficient permission' })
 	async deleteLogo(@Param('number') number: number) {
+		if (typeof number !== 'number' && parseInt(number, 10) != number)
+			throw new BadRequestException(Errors.Generic.FieldInvalid({ i18n: this.i18n, type: Number, field: 'number' }));
+
 		return this.promotionsService.deleteLogo(number);
 	}
 
@@ -110,6 +130,9 @@ export class PromotionsController {
 	@ApiNotFoundResponse({ description: 'Promotion not found' })
 	@ApiUnauthorizedResponse({ description: 'Insufficient permission' })
 	async get(@Param('number') number: number) {
+		if (typeof number !== 'number' && parseInt(number, 10) != number)
+			throw new BadRequestException(Errors.Generic.FieldInvalid({ i18n: this.i18n, type: Number, field: 'number' }));
+
 		return this.promotionsService.findOne(number);
 	}
 
@@ -122,6 +145,9 @@ export class PromotionsController {
 	@ApiNotFoundResponse({ description: 'Promotion not found' })
 	@ApiUnauthorizedResponse({ description: 'Insufficient permission' })
 	async getUsers(@Param('number') number: number) {
+		if (typeof number !== 'number' && parseInt(number, 10) != number)
+			throw new BadRequestException(Errors.Generic.FieldInvalid({ i18n: this.i18n, type: Number, field: 'number' }));
+
 		return this.promotionsService.getUsers(number);
 	}
 }
